requirements.txt
Django==1.8.2
django-autocomplete-light==2.2.2
django-model-utils==2.2
django-reversion==1.8.7
pytz==2015.4
